Almost done
Couple more sections i would like to complete
and once that is done then we can go ahead and
wrap this project up and move onto the next.
I really love react and have had a great time.
I want to update mens and womens sections tomorrow
along with the Home Page, and finish the day with
going through a couple read Mes and working on
battleship along with my portfolio